Fixing the code of conduct pointer
When the v3 branch was started the code of conduct pointed to the
Kubernetes one which pointed to the CNCF. Helm moved to be a CNCF
project and the v2 code of conduct was pointed directly at the
CNCF one. This was missed on the v3 branch. This updates the
pointer.

Signed-off-by: Matt Farina <matt@mattfarina.com>