Fixes a few bugs related to tls config when installing charts:
    When installing via relative path, tls config for the selected
    repository was not being set.

    The --ca-file flag was not being passed when constructing the
    downloader.

    Setting tls config was not checking for zero value in repo
    config, causing flag to get overwritten with empty string.

There's still a few oddities here. I would expect that the flag
passed in on the command line would override the repo config, but
that's not currently the case. Also, we always set the cert, key
and ca files as a trio, when they should be set individually
depending on combination of flags / repo config.

Signed-off-by: James McElwain jmcelwain@gmail.com